refactor(vscode): reorganize webview component folder structure
Move loose root-level components into feature folders and extract
cross-feature reusable components into a shared/ layer:

- components/profile/ — ProfileView + DeviceAuthCard (was at root)
- components/settings/Settings.tsx — moved from root into settings/
- components/shared/ — ModelSelector, ModeSwitcher, ThinkingSelector,
  WorkingIndicator, model-selector-utils (extracted from chat/)

Fixes the cross-folder coupling where settings/ProvidersTab and
agent-manager/ imported from chat/ for reusable selector components.
Update all import paths in App.tsx, PromptInput, MessageList,
ProvidersTab, AgentManagerApp, MultiModelSelector, and settings/index.ts.
