+++ packages/kilo-vscode/webview-ui/src/App.tsx
@@ -8,8 +8,8 @@ import { Code } from "@kilocode/kilo-ui/code"
 import { Diff } from "@kilocode/kilo-ui/diff"
 import { DataProvider } from "@kilocode/kilo-ui/context/data"
 import { Toast } from "@kilocode/kilo-ui/toast"
-import Settings from "./components/Settings"
-import ProfileView from "./components/ProfileView"
+import Settings from "./components/settings/Settings"
+import ProfileView from "./components/profile/ProfileView"
 import { VSCodeProvider, useVSCode } from "./context/vscode"
 import { ServerProvider, useServer } from "./context/server"
 import { ProviderProvider, useProvider } from "./context/provider"
