+++ packages/kilo-vscode/webview-ui/src/components/chat/PromptInput.tsx
@@ -11,9 +11,9 @@ import { useSession } from "../../context/session"
 import { useServer } from "../../context/server"
 import { useLanguage } from "../../context/language"
 import { useVSCode } from "../../context/vscode"
-import { ModelSelector } from "./ModelSelector"
-import { ModeSwitcher } from "./ModeSwitcher"
-import { ThinkingSelector } from "./ThinkingSelector"
+import { ModelSelector } from "../shared/ModelSelector"
+import { ModeSwitcher } from "../shared/ModeSwitcher"
+import { ThinkingSelector } from "../shared/ThinkingSelector"
 import { useFileMention } from "../../hooks/useFileMention"
 import { useImageAttachments } from "../../hooks/useImageAttachments"
 import { fileName, dirName, buildHighlightSegments } from "./prompt-input-utils"
