+++ packages/kilo-vscode/webview-ui/src/components/settings/ProvidersTab.tsx
@@ -6,7 +6,7 @@ import { IconButton } from "@kilocode/kilo-ui/icon-button"
 import { useConfig } from "../../context/config"
 import { useProvider } from "../../context/provider"
 import { useLanguage } from "../../context/language"
-import { ModelSelectorBase } from "../chat/ModelSelector"
+import { ModelSelectorBase } from "../shared/ModelSelector"
 import type { ModelSelection } from "../../types/messages"
 import SettingsRow from "./SettingsRow"
