+++ packages/kilo-vscode/webview-ui/src/components/settings/Settings.tsx
@@ -3,22 +3,22 @@ import { Button } from "@kilocode/kilo-ui/button"
 import { Icon } from "@kilocode/kilo-ui/icon"
 import { Tabs } from "@kilocode/kilo-ui/tabs"
 import { Tooltip } from "@kilocode/kilo-ui/tooltip"
-import { useLanguage } from "../context/language"
-import ProvidersTab from "./settings/ProvidersTab"
-import AgentBehaviourTab from "./settings/AgentBehaviourTab"
-import AutoApproveTab from "./settings/AutoApproveTab"
-import BrowserTab from "./settings/BrowserTab"
-import CheckpointsTab from "./settings/CheckpointsTab"
-import DisplayTab from "./settings/DisplayTab"
-import AutocompleteTab from "./settings/AutocompleteTab"
-import NotificationsTab from "./settings/NotificationsTab"
-import ContextTab from "./settings/ContextTab"
-import TerminalTab from "./settings/TerminalTab"
-import PromptsTab from "./settings/PromptsTab"
-import ExperimentalTab from "./settings/ExperimentalTab"
-import LanguageTab from "./settings/LanguageTab"
-import AboutKiloCodeTab from "./settings/AboutKiloCodeTab"
-import { useServer } from "../context/server"
+import { useLanguage } from "../../context/language"
+import ProvidersTab from "./ProvidersTab"
+import AgentBehaviourTab from "./AgentBehaviourTab"
+import AutoApproveTab from "./AutoApproveTab"
+import BrowserTab from "./BrowserTab"
+import CheckpointsTab from "./CheckpointsTab"
+import DisplayTab from "./DisplayTab"
+import AutocompleteTab from "./AutocompleteTab"
+import NotificationsTab from "./NotificationsTab"
+import ContextTab from "./ContextTab"
+import TerminalTab from "./TerminalTab"
+import PromptsTab from "./PromptsTab"
+import ExperimentalTab from "./ExperimentalTab"
+import LanguageTab from "./LanguageTab"
+import AboutKiloCodeTab from "./AboutKiloCodeTab"
+import { useServer } from "../../context/server"
 
 export interface SettingsProps {
   onBack?: () => void
