+++ packages/kilo-vscode/webview-ui/src/components/shared/ModeSwitcher.tsx
