+++ packages/kilo-vscode/webview-ui/src/components/shared/ThinkingSelector.tsx
