+++ packages/kilo-vscode/webview-ui/src/components/shared/WorkingIndicator.tsx
