+++ packages/kilo-vscode/webview-ui/src/components/shared/model-selector-utils.ts
